+++ CHANGELOG.md
@@ -8,9 +8,11 @@
     - Fix a seed-chaining bug.
     - Fix a target sequence extracting bug.
     - Add output of query start and end positions.
-- New command: `lexicmap utils gen-masks`.
 - `lexicmap utils kmers`:
     - Fix a little bug, wrong number of k-mers for the second k-mer in each k-mer pair.
+- New commands:
+    - `lexicmap utils gen-masks` for generate masks from the top N largest genomes.
+    - `lexicmap utils seed-pos` for extracting seed positions via reference names.
 
 ### v0.2.0 - 2024-02-02